# Try to remove redundant data move in radix sort