--- a/src/AggregateFunctions/QuantileTDigest.h
+++ b/src/AggregateFunctions/QuantileTDigest.h
@@ -103,6 +103,7 @@ class QuantileTDigest
     struct RadixSortTraits
     {
         using Element = Centroid;
+        using Index = Element;
         using Key = Value;
         using CountType = UInt32;
         using KeyBits = UInt32;
@@ -114,6 +115,7 @@ class QuantileTDigest
 
         /// The function to get the key from an array element.
         static Key & extractKey(Element & elem) { return elem.mean; }
+        static Index & extractIndex(Element & elem) { return elem; }
     };
 
     /** Adds a centroid `c` to the digest