@champtar 2020-05-17
Hi @g-arslan 
I'm curious what numbers you get before/after your patch for this test case:
https://github.com/ClickHouse/ClickHouse/issues/2668#issuecomment-598313960