@g-arslan 2020-05-17
Hi @champtar 
Actually it shouldn't affect QuantileTDigest, the only case this patch may improve something is when you need permutation of sorted data