@alexey-milovidov 2020-05-17
@champtar Yes, it's completely independent - QuantileTDigest has some simple flaw in implementation (independent of particular choice of sort algorithm), we should dig into it.