@champtar 2020-05-17
I saw this PR is touching QuantileTDigest.h and RadixSort.h so I remembered my old ticket, but my intuition was wrong, thanks for the clarification.