@alexey-milovidov 2020-05-23
Yes, it is profitable.

```
clickhouse-benchmark <<< "SELECT rand64() AS x FROM numbers(1000000) ORDER BY x"
```

New:
```
Queries executed: 710.

localhost:9000, queries 710, QPS: 11.472, RPS: 12029649.997, MiB/s: 91.779, result RPS: 11472368.238, result MiB/s: 87.527.

0.000%          0.079 sec.
10.000%         0.084 sec.
20.000%         0.085 sec.
30.000%         0.086 sec.
40.000%         0.086 sec.
50.000%         0.087 sec.
60.000%         0.088 sec.
70.000%         0.088 sec.
80.000%         0.089 sec.
90.000%         0.090 sec.
95.000%         0.092 sec.
99.000%         0.094 sec.
99.900%         0.096 sec.
99.990%         0.096 sec.
```

Old:
```
Queries executed: 720.

localhost:9000, queries 720, QPS: 10.941, RPS: 11471997.575, MiB/s: 87.524, result RPS: 10940549.445, result MiB/s: 83.470.

0.000%          0.082 sec.
10.000%         0.088 sec.
20.000%         0.089 sec.
30.000%         0.090 sec.
40.000%         0.091 sec.
50.000%         0.092 sec.
60.000%         0.092 sec.
70.000%         0.093 sec.
80.000%         0.094 sec.
90.000%         0.095 sec.
95.000%         0.096 sec.
99.000%         0.098 sec.
99.900%         0.101 sec.
99.990%         0.101 sec.
```

About 6% improvement.
I will add performance test.